Move agent actions to Node
